requirements.txt
APScheduler==3.6.3
astroid==2.6.5
cachetools==4.2.2
certifi==2021.5.30
flake8==3.9.2
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.9.5
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
wrapt==1.12.1
